Move constants outside of lib folder && Add color to command error && Messages to plugin
