+++ src/api/index.ts
@@ -11,6 +11,6 @@ export * from './locator'
 export * from './modules'
 export * from './table'
 export * from './verbose'
-export * from './lib'
+export * from './constants'
 export { BatchStream, WorkspaceResponse } from './typings'
 export { default as logger } from './logger'
