+++ src/api/modules/workspace/create.ts
@@ -105,7 +105,7 @@ export const workspaceCreator: WorkspaceCreator = async ({
   } catch (err) {
     if (err.response?.data.code === 'ArgumentNull') {
       throw createFlowIssueError(
-        `You need to pick a name for the new workspace. Run vtex worspace create [WORKSPACENAME]. You can name it using your name with a number, for example vtex workspace create johndoe01`
+        `You need to pick a name for the new workspace. Run ${chalk.yellow('vtex workspace create [WORKSPACENAME]')}. You can name it using your name with a number, for example vtex workspace create johndoe01`
       )
     }
