+++ src/lib/auth/AuthProviders/OAuthAuthenticator/index.ts
@@ -7,7 +7,7 @@ import { join } from 'path'
 import logger from '../../../../api/logger'
 import { VTEXID } from '../../../../api/clients/IOClients/external/VTEXID'
 import { storeUrl } from '../../../../api/storeUrl'
-import { ColorifyConstants } from '../../../../api/lib/constants/Colors'
+import { ColorifyConstants } from '../../../../api/constants/Colors'
 import { randomCryptoString } from '../../../utils/randomCryptoString'
 import { spawnUnblockingChildProcess } from '../../../utils/spawnUnblockingChildProcess'
 import { AuthProviderBase } from '../AuthProviderBase'
